better place for users to silence model card
